refactor: separate energy mix composition from water and carbon footprints
Before, we were repeating the energy mix composition in the "carbon_emission_mix" and "water_footprint_mix" host properties to be parsed by the plugin. Now, we separated the energy mix from the respective footprints, removing the duplication and the risk of define two energy mix composed by different percentages.

Signed-off-by: Victor Lima <lima.p.victor@outlook.com>