feat: add public methods to update energy grid attributes individually
Add public methods to be able to update the energy grid composition, the carbon footprint of a/some source(s) and the water footprint of a/some source(s) individually, each in each call.
Optmize the update method, storing the current weighted footprint of carbon (or water), since this value is constant while the energy mix doesn't change.

Signed-off-by: Victor Lima < lima.p.victor@outlook.com>